switched to tito standard builder to speed up build process.
As we are using containers now anyway (and hence a defined env) there
is no need anymore to use the standard env provided by mock.

Furthermore, a new approach is tested: as tito can not deal with git
submodules, vendor code is cloned at build time of the RPM, including all
submodules.